Update patch set 1
Patch Set 1:

When do you plan on adding lte? The recoveries are different.

https://github.com/Shabbypenguin/android_device_motorola_peregrine/commit/b7acb2337f8417ceeb5108233a22c4920c99abc0
Sdcard1 directories also missing from this tree.
I'd prefer adding the peregrine to another device tree and mimic the falcon's current setup, with the addition of a few more things such as gyro being that the xt1042 (cdma) is rumored to be released soon.

I'd also prefer adding a defconfig for the 8926 with this: CONFIG_MMI_PEREGRINE_DTB=y enabled.

Patch-set: 1