Update patch set 1
Patch Set 1:

If your idea of getting the peregrine unified with the falcon is still going to happen, you may want to figure how to workaround this: https://github.com/CyanogenMod/android_kernel_motorola_msm8226/commit/a3760ecf75920495571d06540d6900b2ac97849c

That won't break boot for the peregrine if it is reverted or the flag is added to the defconfig, but if the commit were to stay then the falcon may have an issue booting.

Also, my testing has resulted in weaker radio reception on the falcon's ril libs whereas the peregrines is much stronger. 

I see you just reverted the peregrine add to the falcon tree.

Patch-set: 1