Update patch set 1
Patch Set 1:

ShabbyPenguin and I were nearly finished with the original bring-up, minus the issues with device-only gps.

https://github.com/OUDhs/proprietary_vendor_motorola/commits/update
That's with your un-commonization change to vendor

https://github.com/OUDhs/android_device_motorola_peregrine/commits/rebase
Uncommonized device

Most of the libs are identical from falcon to peregrine, however there is a weak signal on the peregrine on cm vs. stock. The only difference I found was the sensors.msm8226.so breaking the sensors for the peregrine.
I was mistaken into believing the peregrine libs increased the strength of the signal (placebo).

Maybe it would be best to keep them separate until the Moto E is added to the common.

I'll be pushing a change to add a defconfig for the msm8926.

Patch-set: 1