Update patch set 1
Patch Set 1:

This is enabled only in countries call recording is legal.
The one where it is illegal, this is disabled by the country specific mcc codes.

If your asking if this works on falcon, then yes it does.
I have test it on my device with no issues.

Apparently, from the last comment on this change (http://review.cyanogenmod.org/#/c/75492/)
call recording is supposed to be enabled per device the supports it.
Hence this commit.

Patch-set: 1