Update patch set 2
Patch Set 2:

This is the problem with all powerful tools, 
eg: internet, knife, 
It's completely on the user on how it is used.
In developing countries with high corruption this can be a very helpful tool but of course someone could unjustly use it to breach your privacy.

People do good and bad with ordinary things everyday, taking away one thing isn't gonna stop them.

That is why there are laws that incriminate the doer and not the builder. If the creator was blamed for everything we wouldn't have anything.

This is the basic good v/s bad dilemma, the thing is we all have to be hopeful that people will choose good over bad.

Patch-set: 2
Reviewer: Shaleen Jain <10037@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0