Update patch set 2
Patch Set 2: -Code-Review

I understand, and I don't think I need this feature myself. Only noticed it because it's in the falcon device tree. My only concern was that for consistency, it should be allowed on this device when it's allowed on other devices. However, I did not check if that is really the case...

So feel free to abandon this change.

Patch-set: 2