Update patch set 1
Patch Set 1:

Yeah no, the problem is sony doesn't have a separate /firmware partition, and since the bootloader / tz / etc differs, firmware differs too (signature?), and shipping them both in the same build would lead to a ~20% increase in build size.

  
49M     
vendor/sony/togari_gpe/proprietary

We do ship firmware for the non-LTE versions and copy that with a script during installation, but that's just 3 files and a small difference.

Patch-set: 1
Label: Code-Review=0