Update patch set 1
Patch Set 1: Code-Review+1

Currently it returns 'TYPE=(fstype)' when it just needs to return fstype.  I was going to do another cut with -d '"' -f2 but was scared it would freak out with " being a deliminator in the middle of C code and think the " was the end of the syntax.

If the fstype is only 3 characters instead of 4 it will return with an " at the end, but we only care if it is ext4 which is 4 characters.

If you think " as a deliminator won't freak it out feel free to change it to /cut -d '"' -f2

Patch-set: 1