Update patch set 1
Patch Set 1:

(2 comments)

As discussed in http://review.cyanogenmod.org/#/c/65055/ this doesn't need additional cut because we're using strstr to compare the strings (partial match is enough).

This is still "broken" on f2fs partitions because the third field cut returns is incorrect (UUID on f2fs vs. TYPE on ext4). Not that it really matters, though. Suggestions inline.

Patch-set: 1