Update patch set 1
Patch Set 1:

I have this enabled for a while now, and i noticed that it's a bit bugged on falcon too, sometimes it works as it should, but after 3 or 4 times in a row that the prox is covered and you try to unlock the device, it just unlocks without checking the proximity sensor.

This is how it should work:

<6>[  163.074318,0] ct406_work_prox_start: Prox mode start
<4>[  163.075382,0] done.
<6>[  163.075719,0] akm8963 2-000c: akm8963_resume: Resume
<6>[  163.140480,0] PM: suspend exit 2015-08-19 07:57:21.817553933 UTC
<6>[  163.161743,0] ct406_prox_mode_covered: Prox mode covered
<6>[  163.163943,0] ct406_disable_prox: Prox mode stop

This is how it works sometimes, as you can see, it doesn't do the check:

<6>[  164.078679,0] ct406_work_prox_start: Prox mode start
<6>[  164.183025,0] ct406_disable_prox: Prox mode stop

and then unlocks.

If you plug it via USB (charger or PC), is even worse, it unlocks always the 2nd time.

Also, i got some spams:

<3>[  175.600461,0] akm8963 2-000c: AKECS_ioctl: Get Open returns error (-512).
<3>[  175.600877,1] Freezing of tasks aborted after 0.001 seconds (50 tasks refusing to freeze, wq_busy=0):
<4>[  175.601099,1] Restarting tasks ... done.

Other than that, there are no relevant bugs occurring.

Patch-set: 1