Update patch set 1
Patch Set 1:

Yes, I noticed that too. The problem is not our sensor/sensor driver, but the feature itself. The sensor is not activated at all the second time.

The problem seems to happen when there's some background activity. Try to keep logcat running or start listening to music. The screen is kept off only for the first power button press. After that one press, any power button press can turn the screen on, independently on the interval.

Patch-set: 1