Update patch set 1
Patch Set 1:

(2 comments)

Fingerprints:
All the fingerprints are outdated. I think you could remove the fingerprint update from this commit and create (not necessarily you) a new commit that update all the fingerprints.


Converted XT1033:
Even though it's very unlikely, XT1033 users might convert their data partition to ext4 and this would break things for them, but again, I'm not sure if this actually affects anyone.

My main concern here is that I don't know what might be the consequences of converting an XT1033 and I don't know if we should "support" that. It's obvious that someone who did that knew what he was doing and should be aware of all the possible consequences, but I'd like to hear Ethan's opinion on this.

Patch-set: 1